Refactor NPC auto attack
